Re-Publish 0.3.4 from Meteor 1.2.1 (no functionality changes).
This should ensure that the `.versions` file has the appropriate
versions of core packages shipped with it rather than the much newer
versions which are currently included.  Version 1.2.1 was the version
specified for the `versionsFrom` in the `package.json`, which is why I
picked that version.

This ensures that this package will actually work with Meteor 1.2.1,
which is important for some apps!